% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_categerical_vars}
\alias{get_categerical_vars}
\title{Get names of categorical variables from a data frame}
\usage{
get_categerical_vars(data, exclude = NULL, include_char = FALSE)
}
\arguments{
\item{data}{A data frame to extract categorical variable names from.}

\item{exclude}{A character vector of variable names to exclude from the results.
Default is \code{NULL} (no exclusions).}

\item{include_char}{Logical. If \code{TRUE}, character variables are included in
addition to factor variables. Default is \code{FALSE}.}
}
\value{
A character vector of variable names.
}
\description{
Extracts the names of factor variables from a data frame. Optionally includes
character variables and allows exclusion of specific variable names.
}
\examples{
# Get only factor variables
get_categerical_vars(iris)

# Include character variables
get_categerical_vars(mtcars, include_char = TRUE)

# Exclude specific variables
get_categerical_vars(iris, exclude = c("Species"))

}
