% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compose_spreadsheet.R
\name{mark_percentage_cols}
\alias{mark_percentage_cols}
\title{Mark columns as percentage class}
\usage{
mark_percentage_cols(spread)
}
\arguments{
\item{spread}{A data.table with frequency data}
}
\value{
The modified data.table with percentage class on relevant columns
}
\description{
Mark columns as percentage class
}
\keyword{internal}
