% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{check_input}
\alias{check_input}
\title{Check Input Data and Variables}
\usage{
check_input(data, vars, group)
}
\arguments{
\item{data}{A data frame containing the variables to check.}

\item{vars}{A character vector of variable names that must exist in \code{data}
and be factors.}

\item{group}{A character vector of grouping variable names that must exist
in \code{data} and be factors.}
}
\value{
Invisibly returns \code{TRUE} if all checks pass.
}
\description{
Validates that the input data is a data frame and that specified variable
names exist in the data and are factors.
}
\details{
The function performs the following checks:
\enumerate{
\item \code{data} is a data frame
\item All variables in \code{vars} exist in \code{data}
\item All variables in \code{vars} are factors
\item All variables in \code{group} exist in \code{data}
\item All variables in \code{group} are factors
}

If any check fails, the function stops with an informative error message.
}
\examples{
\dontrun{
df <- data.frame(
  x = factor(c("a", "b", "c")),
  y = factor(c("d", "e", "f")),
  z = factor(c("g", "h", "i"))
)
check_input(df, vars = c("x", "y"), group = "z")
}

}
